backend/requirements.txt
# Backend Dependencies for Render.com Deployment
# Generated from pyproject.toml

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database & ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Authentication & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# API & Validation
python-multipart>=0.0.6
email-validator>=2.1.0

# AI & MCP Integration
openai>=1.12.0
mcp>=0.1.0

# Logging & Utilities
python-json-logger>=2.0.7
thefuzz>=0.20.0
python-dateutil>=2.8.2
parsedatetime>=2.6

# Phase 3: Robust AI Assistant Dependencies
dateparser==1.2.0       # Natural language date parsing (more powerful than parsedatetime)
rapidfuzz==3.6.0        # Fast fuzzy string matching (10-20x faster than thefuzz)
pytest-mock==3.12.0     # Mocking framework for tests

# Load Testing
locust>=2.20.0          # Load testing framework for 100 concurrent users test
